wip - fixing tests
