+++ app/jenkins-for-jira-ui/src/components/ConnectionPanel/ConnectionPanel.test.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import {
 	act, fireEvent, render, screen, waitFor
 } from '@testing-library/react';
+import { invoke } from '@forge/bridge';
 import { ConnectionPanelTop } from './ConnectionPanelTop';
 import { ConnectedState } from '../StatusLabel/StatusLabel';
 import { addConnectedState, ConnectionPanel } from './ConnectionPanel';
 import { EventType, JenkinsServer } from '../../../../src/common/types';
 import * as getAllJenkinsServersModule from '../../api/getAllJenkinsServers';
-import { ConnectionPanelMain } from './ConnectionPanelMain';
 
 const servers: JenkinsServer[] = [
 	{
@@ -290,226 +290,6 @@ describe('Connection Panel Suite', () => {
 				);
 
 				expect(screen.queryByText('IP address:', { exact: false })).not.toBeInTheDocument();
-			});
-		});
-
-		describe('Dropdown menu items', () => {
-			// TODO - add test for Rename - will be done when I build the new server name screen
-
-			// TODO - add test for Connection settings -  will be done when I build the new set up Jenkins screen
-
-			test.only('should handle server disconnection and refreshing servers correctly', async () => {
-				jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce(servers);
-
-				render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={jest.fn()}/>);
-
-				await waitFor(() => {
-					expect(screen.getByText(servers[0].name)).toBeInTheDocument();
-					expect(screen.getByText(servers[1].name)).toBeInTheDocument();
-				});
-
-				const dropdownButton = screen.getByTestId(`dropdown-menu-${servers[1].name}`);
-				fireEvent.click(dropdownButton);
-				fireEvent.click(screen.getByText('Disconnect')); // dropdown item disconnect
-				fireEvent.click(screen.getByText('Disconnect')); // modal button disconnect
-
-				await waitFor(() => {
-					expect(screen.getByText(servers[0].name)).toBeInTheDocument();
-					expect(screen.queryByText(servers[1].name)).not.toBeInTheDocument();
-				});
-			});
-		});
-	});
-
-	describe('Connection Panel Main', () => {
-		const setJenkinsServers = jest.fn();
-		test('should render panel content for PENDING server', async () => {
-			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[7]]);
-
-			render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-			await waitFor(() => {
-				expect(screen.getByText('Connection pending')).toBeInTheDocument();
-			});
-		});
-
-		test('should render panel content for DUPLICATE server', async () => {
-			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[5], servers[6]]);
-
-			render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-			await waitFor(() => {
-				expect(screen.getByText('Duplicate server')).toBeInTheDocument();
-			});
-		});
-
-		test('should render panel content for CONNECTED server without pipeline data', async () => {
-			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[1]]);
-
-			render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-			await act(async () => {
-				await waitFor(() => {
-					expect(screen.getByText('No data received')).toBeInTheDocument();
-					expect(screen.queryByText('Pipeline')).not.toBeInTheDocument();
-					expect(screen.queryByText('Event')).not.toBeInTheDocument();
-					expect(screen.queryByText('Received')).not.toBeInTheDocument();
-				});
-			});
-		});
-
-		test('should render panel content for CONNECTED server with pipeline data', async () => {
-			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[5]]);
-
-			render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-			await waitFor(() => {
-				expect(screen.queryByText('No data received')).not.toBeInTheDocument();
-				expect(screen.getByText('Pipeline')).toBeInTheDocument();
-				expect(screen.getByText('Event')).toBeInTheDocument();
-				expect(screen.getByText('Received')).toBeInTheDocument();
-			});
-		});
-
-		test('should handle server deletion correctly for DUPLICATE SERVERS', async () => {
-			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce(servers);
-
-			render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-			await waitFor(() => {
-				// Both have IP address 10.10.10.10
-				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
-				expect(screen.getByText(servers[2].name)).toBeInTheDocument();
-			});
-
-			// Confirm server that isn't a duplicate does not have a delete button
-			expect(screen.queryByTestId(`delete-button-${servers[0].name}`)).not.toBeInTheDocument();
-
-			const deleteButton = screen.getByTestId(`delete-button-${servers[2].name}`);
-			fireEvent.click(deleteButton);
-
-			await waitFor(() => {
-				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
-				expect(screen.queryByText(servers[2].name)).not.toBeInTheDocument();
-			});
-		});
-
-		describe('Setup guide tab', () => {
-			test('should render SetUpGuide component when there is pluginConfig data for a CONNECTED server', async () => {
-				const server = {
-					name: 'server with plugin config',
-					uuid: '56046af9-d0eb-4efb-8896-ed182ende',
-					pluginConfig: {
-						ipAddress: '10.10.10.11',
-						lastUpdatedOn: new Date()
-					},
-					pipelines: [
-						{
-							name: '#74315',
-							lastEventType: EventType.DEPLOYMENT,
-							lastEventStatus: 'in_progress' as const,
-							lastEventDate: new Date()
-						}
-					]
-				};
-
-				jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([server]);
-
-				render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-				await waitFor(() => {
-					expect(screen.getByText(server.name)).toBeInTheDocument();
-
-					fireEvent.click(screen.getByText('Set up guide'));
-				});
-
-				await waitFor(() => {
-					const setUpGuideText =
-						screen.getByText('To receive build and deployment data from this server:');
-					expect(setUpGuideText).toBeInTheDocument();
-				});
-			});
-
-			test('should render UpdateAvailable component when there is no pluginConfig data for a CONNECTED server', async () => {
-				const server = {
-					name: 'server with no plugin config',
-					uuid: '56046af9-d0eb-4efb-8896-ed182ende',
-					pluginConfig: undefined,
-					pipelines: [
-						{
-							name: '#74315',
-							lastEventType: EventType.DEPLOYMENT,
-							lastEventStatus: 'successful' as const,
-							lastEventDate: new Date()
-						}
-					]
-				};
-
-				jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([server]);
-
-				render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-				await waitFor(() => {
-					expect(screen.getByText(server.name)).toBeInTheDocument();
-					fireEvent.click(screen.getByText('Set up guide'));
-				});
-
-				await waitFor(() => {
-					const updateAvailableText =
-						screen.getByText('This server is connected to Jira and sending data, but is using an outdated Atlassian Software Cloud plugin.');
-					expect(updateAvailableText).toBeInTheDocument();
-				});
-			});
-
-			test('should handle refreshing the panel for a server CONNECTED with pipeline data but no plugin config', async () => {
-				jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[4]]);
-
-				const { rerender } =
-					render(<ConnectionPanel jenkinsServers={servers} setJenkinsServers={setJenkinsServers} />);
-
-				await waitFor(() => {
-					expect(screen.getByText('CONNECTED')).toBeInTheDocument();
-					expect(screen.getByText('Pipeline')).toBeInTheDocument();
-					expect(screen.getByText('Event')).toBeInTheDocument();
-					expect(screen.getByText('Received')).toBeInTheDocument();
-					expect(screen.queryByText('Refresh')).not.toBeInTheDocument();
-					expect(screen.queryByText('To receive build and deployment data from this server:')).not.toBeInTheDocument();
-				});
-
-				await waitFor(() => {
-					fireEvent.click(screen.getByText('Set up guide'));
-				});
-
-				await waitFor(() => {
-					expect(screen.getByText('Refresh')).toBeInTheDocument();
-					expect(screen.queryByText('To receive build and deployment data from this server:')).not.toBeInTheDocument();
-
-					const updatedServerData = {
-						...servers[1],
-						pluginConfig: {
-							ipAddress: '10.10.10.12',
-							lastUpdatedOn: new Date()
-						}
-					};
-
-					jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([updatedServerData]);
-
-					// Rerender the component with the updated data
-					rerender(<ConnectionPanelMain
-						connectedState={ConnectedState.CONNECTED}
-						jenkinsServer={updatedServerData}
-						refreshServers={jest.fn()}
-					/>);
-				});
-
-				await waitFor(() => {
-					fireEvent.click(screen.getByText('Set up guide'));
-				});
-
-				await waitFor(() => {
-					fireEvent.click(screen.getByText('Refresh'));
-					expect(screen.getByText('To receive build and deployment data from this server:')).toBeInTheDocument();
-				});
 			});
 		});
 	});
