+++ app/jenkins-for-jira-ui/src/components/ServerManagement/ServerManagement.test.tsx
@@ -1,14 +1,164 @@
 import React from 'react';
 import {
-	fireEvent, render, waitFor, screen
+	fireEvent, render, waitFor, screen, act
 } from '@testing-library/react';
 import { getSiteNameFromUrl, ServerManagement } from './ServerManagement';
+import * as getAllJenkinsServersModule from '../../api/getAllJenkinsServers';
+import * as redirectFromGetStartedModule from '../../api/redirectFromGetStarted';
+import * as fetchGlobalPageUrlModule from '../../api/fetchGlobalPageUrl';
+import { EventType, JenkinsServer } from '../../../../src/common/types';
+import {ConnectionPanel} from "../ConnectionPanel/ConnectionPanel";
+import {invoke} from "@forge/bridge";
+
+const servers: JenkinsServer[] = [
+	{
+		name: 'server one',
+		uuid: '56046af9-d0eb-4efb-8896-6c9d0da884fe',
+		pluginConfig: {
+			ipAddress: '10.10.10.10',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: [
+			{
+				name: '#74315',
+				lastEventType: EventType.DEPLOYMENT,
+				lastEventStatus: 'successful',
+				lastEventDate: new Date()
+			},
+			{
+				name: '#1234',
+				lastEventType: EventType.BUILD,
+				lastEventStatus: 'failed',
+				lastEventDate: new Date()
+			}
+		]
+	},
+	{
+		name: 'server two',
+		uuid: '56046af9-d0eb-4efb-8896-jsdfn8234234',
+		pluginConfig: {
+			ipAddress: '10.10.10.11',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: []
+	},
+	{
+		name: 'server three',
+		uuid: '56046af9-d0eb-4efb-8896-ehdf34bhsdf',
+		pluginConfig: {
+			ipAddress: '10.10.10.10',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: []
+	},
+	{
+		name: 'server four',
+		uuid: '56046af9-d0eb-4efb-8896-sjnd893rsd',
+		pluginConfig: {
+			ipAddress: '10.10.10.12',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: [
+			{
+				name: '#3456',
+				lastEventType: EventType.BUILD,
+				lastEventStatus: 'successful',
+				lastEventDate: new Date()
+			}
+		]
+	},
+	{
+		name: 'server five',
+		uuid: '56046af9-d0eb-4efb-8896-ed182ende',
+		pluginConfig: undefined,
+		pipelines: [
+			{
+				name: '#6789',
+				lastEventType: EventType.DEPLOYMENT,
+				lastEventStatus: 'pending',
+				lastEventDate: new Date()
+			}
+		]
+	},
+	{
+		name: 'server six',
+		uuid: '56046af9-d0eb-4efb-8896-hsdbf723rh2r',
+		pluginConfig: {
+			ipAddress: '10.10.10.10',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: [
+			{
+				name: '#3456',
+				lastEventType: EventType.BUILD,
+				lastEventStatus: 'successful',
+				lastEventDate: new Date()
+			}
+		]
+	},
+	{
+		name: 'server seven',
+		uuid: '56046af9-d0eb-4efb-8896-iwer23rjesu',
+		pluginConfig: {
+			ipAddress: '10.10.10.10',
+			lastUpdatedOn: new Date()
+		},
+		pipelines: []
+	},
+	{
+		name: 'server eight',
+		uuid: '56046af9-d0eb-4efb-8896-iwer23rjesu',
+		pluginConfig: undefined,
+		pipelines: []
+	}
+];
 
 document.execCommand = jest.fn();
+jest.mock('../../api/fetchGlobalPageUrl');
+jest.mock('../../api/redirectFromGetStarted');
+
+describe('getSiteNameFromUrlt', () => {
+	test('correctly extracts site name from URL', async () => {
+		const url = 'https://testjira.atlassian.net/jira/apps/blah-blah';
+		const siteName = getSiteNameFromUrl(url);
+		expect(siteName).toEqual('testjira.atlassian.net');
+	});
+});
 
 describe('ServerManagement Component', () => {
-	test('should copy to clipboard when "Copy to clipboard" is clicked', async () => {
+	test('should render loader when the module key is "get-started-page"', async () => {
+		jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[4]]);
+		jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('get-started-page');
+
 		render(<ServerManagement />);
+
+		expect(screen.getByTestId('jenkins-spinner')).toBeInTheDocument();
+	});
+
+	test('should render the ConnectionWizard when there is an unknown module key', async () => {
+		jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[4]]);
+		jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('unknown-page');
+
+		await waitFor(() => render(<ServerManagement />));
+
+		expect(screen.getByTestId('connection-wizard')).toBeInTheDocument();
+	});
+
+	test('should render the ConnectionWizard when there are no servers', async () => {
+		jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([]);
+		jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+		await waitFor(() => render(<ServerManagement />));
+
+		expect(screen.getByTestId('connection-wizard')).toBeInTheDocument();
+	});
+
+	test('should copy to clipboard when "Copy to clipboard" is clicked', async () => {
+		jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[4]]);
+		jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+		await waitFor(() => render(<ServerManagement />));
+
 		fireEvent.click(screen.getByText('Share page'));
 		fireEvent.click(screen.getByText('Copy to clipboard'));
 
@@ -17,7 +167,11 @@ describe('ServerManagement Component', () => {
 	});
 
 	test('should close the share modal when "Close" is clicked', async () => {
-		render(<ServerManagement />);
+		jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce(servers);
+		jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+		await waitFor(() => render(<ServerManagement />));
+
 		fireEvent.click(screen.getByText('Share page'));
 		expect(screen.getByText('Copy to clipboard')).toBeInTheDocument();
 
@@ -30,41 +184,242 @@ describe('ServerManagement Component', () => {
 		});
 	});
 
-	test('correctly extracts site name from URL', () => {
-		const url = 'https://testjira.atlassian.net/jira/apps/blah-blah';
-		const siteName = getSiteNameFromUrl(url);
-		expect(siteName).toEqual('testjira.atlassian.net');
+	describe.skip('Main panel states', () => {
+		test('should render panel content for PENDING server', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[7]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await waitFor(() => {
+				expect(screen.getByText('Connection pending')).toBeInTheDocument();
+			});
+		});
+
+		test('should render panel content for DUPLICATE server', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[0], servers[2]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await waitFor(() => {
+				expect(screen.getByText('Duplicate server')).toBeInTheDocument();
+			});
+		});
+
+		test('should render panel content for CONNECTED server without pipeline data', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[1]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await act(async () => {
+				await waitFor(() => {
+					expect(screen.getByText('No data received')).toBeInTheDocument();
+					expect(screen.queryByText('Pipeline')).not.toBeInTheDocument();
+					expect(screen.queryByText('Event')).not.toBeInTheDocument();
+					expect(screen.queryByText('Received')).not.toBeInTheDocument();
+				});
+			});
+		});
+
+		test('should render panel content for CONNECTED server with pipeline data', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[3]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await waitFor(() => {
+				expect(screen.queryByText('No data received')).not.toBeInTheDocument();
+				expect(screen.getByText('Pipeline')).toBeInTheDocument();
+				expect(screen.getByText('Event')).toBeInTheDocument();
+				expect(screen.getByText('Received')).toBeInTheDocument();
+			});
+		});
+
+		test('should handle refreshing the panel for a server CONNECTED with pipeline data but no plugin config', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[4]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+			jest.spyOn(fetchGlobalPageUrlModule, 'fetchGlobalPageUrl').mockResolvedValueOnce('https://somesite.atlassian.net/blah');
+
+			const { rerender } = await waitFor(() => render(<ServerManagement />));
+
+			await act(async () => {
+				await waitFor(() => {
+					expect(screen.getByText('CONNECTED')).toBeInTheDocument();
+					expect(screen.getByText('Pipeline')).toBeInTheDocument();
+					expect(screen.getByText('Event')).toBeInTheDocument();
+					expect(screen.getByText('Received')).toBeInTheDocument();
+					expect(screen.queryByText('Refresh')).not.toBeInTheDocument();
+					expect(screen.queryByText('To receive build and deployment data from this server:')).not.toBeInTheDocument();
+				});
+			});
+
+			await waitFor(() => {
+				fireEvent.click(screen.getByText('Set up guide'));
+			});
+
+			await waitFor(async () => {
+				expect(screen.getByText('Refresh')).toBeInTheDocument();
+				expect(screen.queryByText('To receive build and deployment data from this server:')).not.toBeInTheDocument();
+
+				const updatedServerData = {
+					...servers[1],
+					pluginConfig: {
+						ipAddress: '10.10.10.12',
+						lastUpdatedOn: new Date()
+					}
+				};
+
+				jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([updatedServerData]);
+
+				await waitFor(() => rerender(<ServerManagement />));
+			});
+
+			await waitFor(() => {
+				fireEvent.click(screen.getByText('Set up guide'));
+			});
+
+			await waitFor(() => {
+				fireEvent.click(screen.getByText('Refresh'));
+				expect(screen.getByText('To receive build and deployment data from this server:')).toBeInTheDocument();
+			});
+		});
+
+		test('should handle server deletion correctly for DUPLICATE SERVERS', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([servers[0], servers[2]]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await waitFor(() => {
+				// Both have IP address 10.10.10.10
+				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
+				expect(screen.getByText(servers[2].name)).toBeInTheDocument();
+			});
+
+			// Confirm server that isn't a duplicate does not have a delete button
+			expect(screen.queryByTestId(`delete-button-${servers[0].name}`)).not.toBeInTheDocument();
+
+			const deleteButton = screen.getByTestId(`delete-button-${servers[2].name}`);
+			fireEvent.click(deleteButton);
+
+			await waitFor(() => {
+				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
+				expect(screen.queryByText(servers[2].name)).not.toBeInTheDocument();
+			});
+		});
 	});
 
-	test('should render loading icon when request is being made for servers', () => {
-		const { getByText } = render(
-			<ServerManagement />
-		);
+	describe.skip('Dropdown menu items', () => {
+		// TODO - add test for Rename - will be done when I build the new server name screen
 
-		expect(getByText('build')).toBeInTheDocument();
+		// TODO - add test for Connection settings -  will be done when I build the new set up Jenkins screen
+
+		test('should handle server disconnection and refreshing servers correctly', async () => {
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce(servers);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => {
+				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
+				expect(screen.getByText(servers[1].name)).toBeInTheDocument();
+			});
+
+			const dropdownButton = screen.getByTestId(`dropdown-menu-${servers[1].name}`);
+			fireEvent.click(dropdownButton);
+
+			await waitFor(() => {
+				expect(screen.getByText('Disconnect')).toBeInTheDocument();
+			});
+
+			fireEvent.click(screen.getByText('Disconnect'));
+
+			await waitFor(() => {
+				expect(screen.getByTestId('disconnectModal')).toBeInTheDocument();
+			});
+
+			fireEvent.click(screen.getByText('Disconnect'));
+
+			await act(async () => {
+				expect(invoke).toHaveBeenCalledWith('disconnectJenkinsServer', { uuid: servers[1].uuid });
+				expect(screen.getByText(servers[0].name)).toBeInTheDocument();
+			});
+
+			await act(async (): Promise<void> => {
+				expect(
+					await waitFor(() => screen.queryByText(servers[1].name))
+				).not.toBeInTheDocument();
+			});
+		});
 	});
 
-	test('should render loading icon when request is being made for the moduleKey', () => {
-		const { getByText } = render(
-			<ServerManagement />
-		);
+	describe('Setup guide tab', () => {
+		test('should render SetUpGuide component when there is pluginConfig data for a CONNECTED server', async () => {
+			const server = {
+				name: 'server with plugin config',
+				uuid: '56046af9-d0eb-4efb-8896-ed182ende',
+				pluginConfig: {
+					ipAddress: '10.10.10.11',
+					lastUpdatedOn: new Date()
+				},
+				pipelines: [
+					{
+						name: '#74315',
+						lastEventType: EventType.DEPLOYMENT,
+						lastEventStatus: 'in_progress' as const,
+						lastEventDate: new Date()
+					}
+				]
+			};
 
-		expect(getByText('build')).toBeInTheDocument();
-	});
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([server]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
 
-	test('should render ConnectionWizard when fetchAllJenkinsServers returns no servers', () => {
-		const { getByText } = render(
-			<ServerManagement />
-		);
+			await waitFor(() => render(<ServerManagement />));
 
-		expect(getByText('build')).toBeInTheDocument();
-	});
+			await waitFor(() => {
+				expect(screen.getByText(server.name)).toBeInTheDocument();
 
-	test('should render ConnectionPanel when fetchAllJenkinsServers returns servers', () => {
-		const { getByText } = render(
-			<ServerManagement/>
-		);
+				fireEvent.click(screen.getByText('Set up guide'));
+			});
 
-		expect(getByText('build')).toBeInTheDocument();
+			await waitFor(() => {
+				const setUpGuideText =
+					screen.getByText('To receive build and deployment data from this server:');
+				expect(setUpGuideText).toBeInTheDocument();
+			});
+		});
+
+		test('should render UpdateAvailable component when there is no pluginConfig data for a CONNECTED server', async () => {
+			const server = {
+				name: 'server with no plugin config',
+				uuid: '56046af9-d0eb-4efb-8896-ed182ende',
+				pluginConfig: undefined,
+				pipelines: [
+					{
+						name: '#74315',
+						lastEventType: EventType.DEPLOYMENT,
+						lastEventStatus: 'successful' as const,
+						lastEventDate: new Date()
+					}
+				]
+			};
+
+			jest.spyOn(getAllJenkinsServersModule, 'getAllJenkinsServers').mockResolvedValueOnce([server]);
+			jest.spyOn(redirectFromGetStartedModule, 'redirectFromGetStarted').mockResolvedValueOnce('jenkins-for-jira-ui-admin-page');
+
+			await waitFor(() => render(<ServerManagement />));
+
+			await waitFor(() => {
+				expect(screen.getByText(server.name)).toBeInTheDocument();
+				fireEvent.click(screen.getByText('Set up guide'));
+			});
+
+			await waitFor(() => {
+				const updateAvailableText =
+					screen.getByText('This server is connected to Jira and sending data, but is using an outdated Atlassian Software Cloud plugin.');
+				expect(updateAvailableText).toBeInTheDocument();
+			});
+		});
 	});
 });
